test: DELETE message

+++ server/tests/testHelper.js
@@ -7,17 +7,29 @@ const getMessages = async (api) => {
 };
 
 const initDb = async (model, initialContent = []) => {
+  if (!model) throw new Error('No model given');
   await model.deleteMany({});
   await model.insertMany(initialContent);
 };
 
 const contentInDb = async (model) => {
+  if (!model) throw new Error('No model given');
   const content = await model.find({});
   return content.map((c) => c.toJSON());
 };
+
+const contentCountInDb = async (model) => {
+  if (!model) throw new Error('No model given');
+  const content = await contentInDb(model);
+  return content.length;
+};
+
+const errorResponse = (message = '') => ({ error: message });
 
 module.exports = {
   getMessages,
   initDb,
   contentInDb,
+  contentCountInDb,
+  errorResponse,
 };
